# openapi.yaml
openapi: 3.0.3
info:
  title: Auth Server API
  description: |
    Auth Server trung tâm được xây dựng bằng Rust để quản lý xác thực và phân quyền cho nhiều hệ thống con (App).
    
    ## Features
    - User registration and authentication
    - JWT-based authentication with RS256 algorithm
    - Token refresh mechanism
    - Password reset flow
    - Multi-app support with scoped roles and permissions
    - Role-Based Access Control (RBAC)
    
    ## Authentication
    Protected endpoints require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
  version: 1.0.0
  contact:
    name: Auth Server Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: |
        Create a new user account with email and password.
        
        Requirements: 14.1, 1.1-1.5
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: "user@example.com"
              password: "SecurePassword123!"
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid email format or weak password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_email:
                  summary: Invalid email format
                  value:
                    error: "invalid_email"
                    message: "Invalid email format"
                    status_code: 400
                weak_password:
                  summary: Weak password
                  value:
                    error: "weak_password"
                    message: "Password does not meet requirements"
                    status_code: 400
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "email_exists"
                message: "Email already exists"
                status_code: 409
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: |
        Authenticate user with email and password.
        
        ## Response Types
        - **Without MFA**: Returns JWT access token and refresh token directly
        - **With MFA**: Returns `mfa_required: true` with `mfa_token` - use `/auth/mfa/verify` to complete login
        
        ## Security Features
        - **Rate Limiting**: 5 login attempts per minute per IP+email combination
        - **Account Lockout**: After 5 failed attempts, account is locked for 15 minutes
        - **Audit Logging**: All login attempts (success/failure) are logged
        - **MFA Support**: If user has MFA enabled, requires 2-step verification
        
        Requirements: 14.2, 2.1-2.5
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "user@example.com"
              password: "SecurePassword123!"
      responses:
        '200':
          description: Successfully authenticated or MFA required
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TokenResponse'
                  - $ref: '#/components/schemas/MfaRequiredResponse'
              examples:
                success:
                  summary: Direct login success (no MFA)
                  value:
                    access_token: "eyJhbGciOiJSUzI1NiIs..."
                    refresh_token: "eyJhbGciOiJSUzI1NiIs..."
                    token_type: "Bearer"
                    expires_in: 900
                mfa_required:
                  summary: MFA verification required
                  value:
                    mfa_required: true
                    mfa_token: "550e8400-e29b-41d4-a716-446655440000"
                    available_methods: ["totp"]
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_credentials"
                message: "Invalid credentials"
                status_code: 401
        '403':
          description: User is inactive or account is locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                user_inactive:
                  summary: User is inactive
                  value:
                    error: "user_inactive"
                    message: "User is inactive"
                    status_code: 403
                account_locked:
                  summary: Account is locked due to failed attempts
                  value:
                    error: "account_locked"
                    message: "Account is locked"
                    status_code: 403
                user_banned:
                  summary: User is banned from the app
                  value:
                    error: "user_banned"
                    message: "User is banned from this app"
                    status_code: 403
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "rate_limit_exceeded"
                message: "Rate limit exceeded"
                status_code: 429
          headers:
            Retry-After:
              description: Number of seconds to wait before retrying
              schema:
                type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/mfa/verify:
    post:
      tags:
        - Authentication
      summary: Complete MFA login
      description: |
        Complete the login process by verifying MFA code.
        Called after `/auth/login` returns `mfa_required: true`.
        
        ## Supported Methods
        - **TOTP**: 6-digit code from authenticator app
        - **Backup Code**: One-time use backup code
        
        ## Security Features
        - **Rate Limiting**: 5 attempts per 5 minutes
        - **Token Expiry**: MFA token expires in 5 minutes
      operationId: completeMfaLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteMfaLoginRequest'
            examples:
              totp:
                summary: TOTP code verification
                value:
                  mfa_token: "550e8400-e29b-41d4-a716-446655440000"
                  code: "123456"
                  is_backup_code: false
              backup:
                summary: Backup code verification
                value:
                  mfa_token: "550e8400-e29b-41d4-a716-446655440000"
                  code: "ABCD1234"
                  is_backup_code: true
      responses:
        '200':
          description: MFA verified successfully, tokens returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid MFA code or token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_code:
                  summary: Invalid MFA code
                  value:
                    error: "invalid_mfa_code"
                    message: "Invalid MFA code"
                    status_code: 401
                invalid_token:
                  summary: Invalid or expired MFA token
                  value:
                    error: "invalid_token"
                    message: "Invalid token"
                    status_code: 401
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Get a new access token using a valid refresh token.
        
        Requirements: 14.3, 3.1-3.3
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            example:
              refresh_token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token successfully refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_token:
                  summary: Invalid token
                  value:
                    error: "invalid_token"
                    message: "Invalid token"
                    status_code: 401
                token_expired:
                  summary: Token expired
                  value:
                    error: "token_expired"
                    message: "Token expired"
                    status_code: 401
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Initiate password reset
      description: |
        Request a password reset token. For security, the response is always the same
        regardless of whether the email exists in the system.
        
        Requirements: 14.4, 4.1-4.2
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
            example:
              email: "user@example.com"
      responses:
        '200':
          description: Password reset initiated (response is always the same for security)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "If the email exists, a password reset link has been sent."
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Complete password reset
      description: |
        Reset password using a valid reset token.
        
        Requirements: 14.5, 4.3-4.4
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
            example:
              token: "reset-token-uuid"
              new_password: "NewSecurePassword123!"
      responses:
        '200':
          description: Password successfully reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Password has been reset successfully."
        '401':
          description: Invalid or expired reset token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_token"
                message: "Invalid token"
                status_code: 401
        '400':
          description: Weak password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "weak_password"
                message: "Password does not meet requirements"
                status_code: 400
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify-email:
    post:
      tags:
        - Authentication
      summary: Verify email address
      description: |
        Verify user's email address using the verification token sent via email.
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
            example:
              token: "verification-token-uuid"
      responses:
        '200':
          description: Email successfully verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Email verified successfully"
        '401':
          description: Invalid or expired verification token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/resend-verification:
    post:
      tags:
        - Authentication
      summary: Resend verification email
      description: |
        Request a new verification email. For security, the response is always the same
        regardless of whether the email exists or is already verified.
      operationId: resendVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
            example:
              email: "user@example.com"
      responses:
        '200':
          description: Verification email sent (response is always the same for security)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "If the email exists and is not verified, a verification link has been sent"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: List all users
      description: List all users with pagination. Requires system admin privileges.
      operationId: listAllUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsersResponse'
        '401':
          description: Unauthorized
        '403':
          description: Not a system admin
...
  /admin/users/{user_id}/activate:
    post:
      tags:
        - Admin
      summary: Activate user
      description: Activate a deactivated user. Requires system admin privileges.
      operationId: activateUser
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User activated
        '404':
          description: User not found
        '403':
          description: Not a system admin
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from /auth/login endpoint.
        Token is signed using RS256 algorithm.
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address (must be unique)
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          description: User's password (must meet security requirements)
          example: "SecurePassword123!"
...
    RegisterResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
...